+++ database/schema.sql
@@ -1055,6 +1055,8 @@ CREATE TABLE `users` (
   `nav_collapsed` tinyint(1) NOT NULL DEFAULT 0,
   `events_sort` varchar(8) DEFAULT NULL,
   `alt_emails` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_bin DEFAULT NULL CHECK (json_valid(`alt_emails`)),
+  `privacy_policy_accepted_at` datetime DEFAULT NULL,
+  `privacy_policy_version` varchar(32) DEFAULT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `email` (`email`)
 ) ENGINE=InnoDB AUTO_INCREMENT=66545 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;